Refactor toward test case independence from Galaxy.
